+++ HISTORY.md
@@ -1,4 +1,36 @@
 
+## Next release
+
+### Added:
+
+- argument 'user_filter' to OPH_CLUSTER
+- process associated with OPH_CONCATNC
+- process associated with OPH_CONCATNC2
+- argument 'algorithm' to OPH_CUBESIZE
+- argument 'container_pid' to OPH_DELETECONTAINER
+- argument 'algorithm' to OPH_RANDCUBE
+- process associated with OPH_RANDCUBE2
+- arguments 'subset_filter' and 'offset' to OPH_SET
+
+### Fixed:
+
+- command name associated with OPH_CLUSTER
+
+### Changed:
+
+- argument 'action' of OPH_CLUSTER
+- argument 'operation' of OPH_INTERCUBE
+
+### Removed:
+
+- arguments 'delete_type' and 'hidden' from OPH_DELETECONTAINER
+- arguments 'filesystem', 'ndbms' and 'ndb' from OPH_IMPORTFITS
+- arguments 'filesystem', 'ndbms' and 'ndb' from OPH_IMPORTNC
+- arguments 'filesystem', 'ndbms' and 'ndb' from OPH_IMPORTNC2
+- arguments 'hidden' from OPH_LIST
+- arguments 'filesystem', 'ndbms' and 'ndb' from OPH_RANDCUBE
+- process associated with OPH_RESTORECONTAINER
+
 ## v1.0.0 - 2018-07-27
 
 - Add a specific process for each operator